Split out Authorization transform into file
This commit moves the `Authorization` struct out of the mod.rs file and
into its own file.
